06_spaceship/requirements.txt
grpcio==1.56.2
protobuf==4.24.0
pydantic==2.1.1
